Outlet-Db-Backup-Monitoring-Backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
pyodbc==5.2.0
gunicorn==23.0.0
smbprotocol==1.15.0
PyJWT==2.10.1
bcrypt==4.3.0
